requirements.txt
Flask
Flask-JWT-Extended
Flask-Script
python-dotenv
flask-cors